+++ ansible/roles/standardise/files/sdf/f40_create_stand_database.sql
@@ -3,13 +3,14 @@
  * Purpose: Creates fragmentation company specific tables
  *
  * Author | Date    | Version
+ * Alan   | 10/2024 | BIGINT Version
  * Duncan | 03/2020 | Initial Version
  *
  */
 
 
 /*
- * Drop all tables if they exist (NB Order because of table key constrants) 
+ * Drop all tables if they exist (NB Order because of table key constrants)
  */
 
 DROP TABLE IF EXISTS i_mols_sdf;
@@ -23,6 +24,6 @@ CREATE TABLE i_mols_sdf (
   nonisosmiles TEXT,
   hac SMALLINT,
   cmpd_id TEXT,
-  isomol_id INTEGER,
-  nonisomol_id INTEGER
+  isomol_id bigint,
+  nonisomol_id bigint
 );
